app #212 api: rename location_overview.dart
